requirements.txt
python-telegram-bot==21.6
SQLAlchemy==2.0.*
python-dotenv
fastapi
uvicorn
pydantic
openpyxl
